Moved the rest of the MyBuilds code into the MyBuilds component
