+++ app/components/layout/Navigation/MyBuilds/index.js
@@ -1,11 +1,18 @@
 import React from 'react';
 import Relay from 'react-relay';
+import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
+import classNames from 'classnames';
 
 import PusherStore from '../../../../stores/PusherStore';
 import Button from '../../../shared/Button';
 import Spinner from '../../../shared/Spinner';
+import Dropdown from '../../../shared/Dropdown';
+import Icon from '../../../shared/Icon';
+import Badge from '../../../shared/Badge';
+import CachedStateWrapper from '../../../../lib/CachedStateWrapper';
 
 import Build from './build';
+import DropdownButton from './../dropdown-button';
 
 class BuildsDropdown extends React.Component {
   static propTypes = {
@@ -14,7 +21,27 @@ class BuildsDropdown extends React.Component {
   }
 
   state = {
-    fetching: true
+    fetching: true,
+    showing: false,
+    scheduledBuildsCount: this.props.viewer.scheduledBuilds ? this.props.viewer.scheduledBuilds.count : 0,
+    runningBuildsCount: this.props.viewer.runningBuilds ? this.props.viewer.runningBuilds.count : 0
+  }
+
+  componentWillMount() {
+    const initialState = {};
+    const cachedState = this.getCachedState();
+
+    if (!this.props.viewer.scheduledBuilds) {
+      initialState.scheduledBuildsCount = cachedState.scheduledBuildsCount || 0;
+    }
+
+    if (!this.props.viewer.runningBuilds) {
+      initialState.runningBuildsCount = cachedState.runningBuildsCount || 0;
+    }
+
+    if (Object.keys(initialState).length) {
+      this.setState(initialState);
+    }
   }
 
   componentDidMount() {
@@ -32,7 +59,45 @@ class BuildsDropdown extends React.Component {
     PusherStore.off("user_stats:change", this.handlePusherWebsocketEvent);
   }
 
+  componentWillReceiveProps = (nextProps) => {
+    if (nextProps.viewer.scheduledBuilds || nextProps.viewer.runningBuilds) {
+      this.setCachedState({
+        scheduledBuildsCount: nextProps.viewer.scheduledBuilds.count,
+        runningBuildsCount: nextProps.viewer.runningBuilds.count
+      });
+    }
+  };
+
   render() {
+    const buildsCount = this.state.runningBuildsCount + this.state.scheduledBuildsCount;
+    let badge;
+
+    if (buildsCount) {
+      badge = (
+        <Badge className={classNames("hover-lime-child", { "bg-lime": this.state.showingDropdown })}>
+          {buildsCount}
+        </Badge>
+      );
+    }
+
+    return (
+      <Dropdown width={320} className="flex" onToggle={this.handleBuildsDropdownToggle}>
+        <DropdownButton className={classNames("py0", { "lime": this.state.showingDropdown })}>
+          {'My Builds '}
+          <div className="xs-hide">
+            <ReactCSSTransitionGroup transitionName="transition-appear-pop" transitionEnterTimeout={200} transitionLeaveTimeout={200}>
+              {badge}
+            </ReactCSSTransitionGroup>
+          </div>
+          <Icon icon="down-triangle" style={{ width: 7, height: 7, marginLeft: '.5em' }} />
+        </DropdownButton>
+
+        {this.renderDropdown()}
+      </Dropdown>
+    );
+  }
+
+  renderDropdown() {
     if (this.state.fetching) {
       return this.renderSpinner();
     } else if (this.props.viewer.user.builds.edges.length > 0) {
@@ -72,6 +137,10 @@ class BuildsDropdown extends React.Component {
     );
   }
 
+  handleBuildsDropdownToggle = (visible) => {
+    this.setState({ showingDropdown: visible });
+  };
+
   handlePusherWebsocketEvent = (payload) => {
     // Only do a relay update of the builds count changes
     if (this._buildsCount !== payload.buildsCount) {
@@ -80,16 +149,26 @@ class BuildsDropdown extends React.Component {
 
     // Save the new builds count
     this._buildsCount = payload.buildsCount;
+
+    this.setCachedState({
+      scheduledBuildsCount: payload.scheduledBuildsCount,
+      runningBuildsCount: payload.runningBuildsCount
+    });
   };
 }
 
-export default Relay.createContainer(BuildsDropdown, {
-  initialVariables: {
-    isMounted: false
-  },
+export default Relay.createContainer(
+  CachedStateWrapper(
+    BuildsDropdown,
+    { validLength: 60 * 60 * 1000 /* 1 hour */ }
+  ),
+  {
+    initialVariables: {
+      isMounted: false
+    },
 
-  fragments: {
-    viewer: () => Relay.QL`
+    fragments: {
+      viewer: () => Relay.QL`
       fragment on Viewer {
         user {
           builds(first: 5) @include(if: $isMounted) {
@@ -101,7 +180,14 @@ export default Relay.createContainer(BuildsDropdown, {
             }
           }
         }
+        runningBuilds: builds(state: BUILD_STATE_RUNNING) @include(if: $isMounted) {
+          count
+        }
+        scheduledBuilds: builds(state: BUILD_STATE_SCHEDULED) @include(if: $isMounted) {
+          count
+        }
       }
     `
+    }
   }
-});
+);
