+++ app/components/layout/Navigation/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import Relay from 'react-relay';
-import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
 import shallowCompare from 'react-addons-shallow-compare';
 import classNames from 'classnames';
 
@@ -13,7 +12,6 @@ import AgentsCount from '../../organization/AgentsCount';
 import NewChangelogsBadge from '../../user/NewChangelogsBadge';
 import permissions from '../../../lib/permissions';
 import PusherStore from '../../../stores/PusherStore';
-import CachedStateWrapper from '../../../lib/CachedStateWrapper';
 
 import NavigationButton from './navigation-button';
 import DropdownButton from './dropdown-button';
@@ -23,86 +21,23 @@ import MyBuilds from './MyBuilds';
 class Navigation extends React.Component {
   static propTypes = {
     organization: React.PropTypes.object,
-    viewer: React.PropTypes.shape({
-      user: React.PropTypes.shape({
-        name: React.PropTypes.string.isRequired,
-        avatar: React.PropTypes.shape({
-          url: React.PropTypes.string.isRequired
-        })
-      }),
-      organizations: React.PropTypes.shape({
-        edges: React.PropTypes.array
-      }),
-      unreadChangelogs: React.PropTypes.shape({
-        count: React.PropTypes.number
-      }),
-      scheduledBuilds: React.PropTypes.shape({
-        count: React.PropTypes.number.isRequired
-      }),
-      runningBuilds: React.PropTypes.shape({
-        count: React.PropTypes.number.isRequired
-      })
-    }),
+    viewer: React.PropTypes.object,
     relay: React.PropTypes.object.isRequired
   };
 
-  componentWillMount() {
-    const initialState = {};
-    const cachedState = this.getCachedState();
-
-    if (!this.props.viewer.scheduledBuilds) {
-      initialState.scheduledBuildsCount = cachedState.scheduledBuildsCount || 0;
-    }
-
-    if (!this.props.viewer.runningBuilds) {
-      initialState.runningBuildsCount = cachedState.runningBuildsCount || 0;
-    }
-
-    if (Object.keys(initialState).length) {
-      this.setState(initialState);
-    }
-  }
 
   componentDidMount() {
     this.props.relay.setVariables({ isMounted: true });
-    PusherStore.on("user_stats:change", this.handlePusherWebsocketEvent);
-  }
-
-  componentWillReceiveProps = (nextProps) => {
-    if (nextProps.viewer.scheduledBuilds || nextProps.viewer.runningBuilds) {
-      this.setCachedState({
-        scheduledBuildsCount: nextProps.viewer.scheduledBuilds.count,
-        runningBuildsCount: nextProps.viewer.runningBuilds.count
-      });
-    }
-  };
-
-  componentWillUnmount() {
-    PusherStore.off("user_stats:change", this.handlePusherWebsocketEvent);
   }
 
   state = {
-    scheduledBuildsCount: this.props.viewer.scheduledBuilds ? this.props.viewer.scheduledBuilds.count : 0,
-    runningBuildsCount: this.props.viewer.runningBuilds ? this.props.viewer.runningBuilds.count : 0,
     showingOrgDropdown: false,
-    showingBuildsDropdown: false,
     showingUserDropdown: false,
     showingSupportDialog: false
   };
 
-  handlePusherWebsocketEvent = (payload) => {
-    this.setCachedState({
-      scheduledBuildsCount: payload.scheduledBuildsCount,
-      runningBuildsCount: payload.runningBuildsCount
-    });
-  };
-
   handleOrgDropdownToggle = (visible) => {
     this.setState({ showingOrgDropdown: visible });
-  };
-
-  handleBuildsDropdownToggle = (visible) => {
-    this.setState({ showingBuildsDropdown: visible });
   };
 
   handleUserDropdownToggle = (visible) => {
@@ -161,31 +96,9 @@ class Navigation extends React.Component {
       return null;
     }
 
-    const buildsCount = this.state.runningBuildsCount + this.state.scheduledBuildsCount;
-    let badge;
-
-    if (buildsCount) {
-      badge = (
-        <Badge className={classNames("hover-lime-child", { "bg-lime": this.state.showingBuildsDropdown })}>
-          {buildsCount}
-        </Badge>
-      );
-    }
-
     return (
-      <Dropdown width={320} className="flex" onToggle={this.handleBuildsDropdownToggle}>
-        <DropdownButton className={classNames("py0", { "lime": this.state.showingBuildsDropdown })}>
-          {'My Builds '}
-          <div className="xs-hide">
-            <ReactCSSTransitionGroup transitionName="transition-appear-pop" transitionEnterTimeout={200} transitionLeaveTimeout={200}>
-              {badge}
-            </ReactCSSTransitionGroup>
-          </div>
-          <Icon icon="down-triangle" style={{ width: 7, height: 7, marginLeft: '.5em' }} />
-        </DropdownButton>
-        <MyBuilds viewer={this.props.viewer} />
-      </Dropdown>
-    );
+      <MyBuilds viewer={this.props.viewer} />
+    )
   }
 
   renderOrganizationsList() {
@@ -345,18 +258,13 @@ class Navigation extends React.Component {
   }
 }
 
-export default Relay.createContainer(
-  CachedStateWrapper(
-    Navigation,
-    { validLength: 60 * 60 * 1000 /* 1 hour */ }
-  ),
-  {
-    initialVariables: {
-      isMounted: false
-    },
+export default Relay.createContainer(Navigation, {
+  initialVariables: {
+    isMounted: false
+  },
 
-    fragments: {
-      organization: () => Relay.QL`
+  fragments: {
+    organization: () => Relay.QL`
         fragment on Organization {
           name
           id
@@ -383,7 +291,7 @@ export default Relay.createContainer(
           }
         }
       `,
-      viewer: () => Relay.QL`
+    viewer: () => Relay.QL`
         fragment on Viewer {
           ${MyBuilds.getFragment('viewer')}
           user {
@@ -403,14 +311,7 @@ export default Relay.createContainer(
           unreadChangelogs: changelogs(read: false) @include(if: $isMounted) {
             count
           }
-          runningBuilds: builds(state: BUILD_STATE_RUNNING) @include(if: $isMounted) {
-            count
-          }
-          scheduledBuilds: builds(state: BUILD_STATE_SCHEDULED) @include(if: $isMounted) {
-            count
-          }
         }
       `
-    }
   }
-);
+});
